+++ js/app.js
@@ -8,6 +8,7 @@ import { initializeSidebarToggle } from "./sidebar-controller.js";
 import Dashboard from "./views/dashboard.js";
 import Items from "./views/items.js";
 import Purchases from "./views/purchases.js";
+import Conversions from "./views/conversions.js";
 import Statistics from "./views/statistics.js";
 import Readme from "./views/readme.js";
 import Settings from "./views/settings.js";
@@ -28,7 +29,7 @@ async function applyVersion(){
   }
 }
 
-[Dashboard, Items, Purchases, Statistics, Readme, Settings].forEach((view) =>
+[Dashboard, Items, Purchases, Conversions, Statistics, Readme, Settings].forEach((view) =>
   Router.register(view),
 );